requirements.txt
Django==4.2
psycopg2==2.9
Pillow==9.5
requests==2.29
django-phonenumber-field[phonenumbers]==7.1.0